// src/app/models/doi/doiModel.ts
export class ListValue {
    value: string;
    viewValue: string;
}
export class BrwoseData {
    name: string;
    file: string;
    uploadedBy: string;
}
export class AttachmentData {
    attachmentType: undefined | string;
    name: undefined | string;
    file: undefined | string;
}

export class MasterEntry {
    schemeName: string;
    schemeId: number;
    schemeNameGuj: string;
    schemeShortname: string;
    nodalOffice: string;
    nodelOfficeId: number;
    schemeStatusId: number;
    claimAmount: number;
    createdDate: string;
    updatedDate: string;
    status: number;
    maximumAgeRange: number;
    minimumAgeRange: number;
    remarks: string;
    referenceNo: string;
    referenceDate: string;
    activeStatus: number;
}
// ...
export class JpaLegalDetails {
    legalEntryId: number;
    inwardDt: string;
    inwardNo: string;
    jpaClaimId: string;
    referenceDt: Date;
    referenceNo: string;
    jpaLegalDtlDTO:  Array<JpaLegal>=[];
}

export class JpaLegal {
    legalDtlsId: number;
    accidentDeathDt: string;
    appealAdmtDt: string;
    appealNo: string;
    appealPaidId: number;
    appealRemarks: string;
    commissionType: string;
    courtDecision: string;
    courtOrderDt: string;
    srno: string;
    district: string;
    inwardNo: number;
    inwardDate: string;
    caseNo: string;
    courtCaseDt: string;
    courtDetail: string;
    claimNumber: string;
    courtCaseNo:string;
    policyNo: string;
    scheme: string;
    schemeType: string;
    applicantName: string;
    nameofDeceasedPerson: string;
    detailOfQuery: string;
    dateOfAccident: string;
    status: string;
    action: string;
    replyVal: string;
    replayDate: string;
    courtDesc: string;
    rpAdmitDate: string;
    revisionPetitionNo: number;
    orderStatus: boolean;
    appealPaid3: number;
    payMode: number;
    paidChequeNo: number;
    ddNo: number;
    paidDate: string;
    paidAmount: number;
    nameOfDese: string;
    doiJpaLegalOthrPaymentDTO :  Array<OtherPayment>=[] ;
    courtTypeId: string;
    deadPersonName: string;
    depPaymntModeId: string;
    depositAmount: number;
    depositDt: string;
    districtCaseNo: string;
    districtId: number;
    orderStatusId: number;
    paidPaymntModeId: number;
    paymentAmount: number;
    paymentDt: string;
    policyNum: string;
    queryDetails: string;
    refPaymntModeId: number;
    refundAmount: number;
    refundDt: string;
    replayDt: string;
    replyDetails:string;
    revPetitionNo: string;
    rpAdmtDt: string;
    schemeId:number;
    cmaNo: string;
    createModeON: string;
    updatedDate: string;
    personName: string;
    paymentModeId:string;
}

export class OtherPayment {
    id: number;
    payMode: string;
    paidRefNo: string;
    paidChequeNo: string;
    ddNo: number;
    paidDate:string;
    paidAmount: number;
    paymentRemarks: string;
    othrPaymntsId: number;
    chequeDt: string;
    chequeNum: number;
    commissionType: string;
    ddNum: number;
    paymentAmount: number;
    paymentModeId: number;
    referenceNo: string;
    remarks: string;

}